refactor(sera-types): unify HarnessError, delete sera-runtime duplicate (sera-8s91/harnerr)
